requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.30.0
attrs==25.3.0
autopep8==2.3.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.104.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.10
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.2
langsmith==0.0.87
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.51.0
packaging==23.2
propcache==0.3.2
pycodestyle==2.14.0
pydantic==2.5.0
pydantic_core==2.14.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
redis==5.0.1
regex==2025.9.1
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.27.0
tenacity==8.5.0
tiktoken==0.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
